Fix menu on input values, debt table input values, validate category delete

+++ src/main/java/com/cafe/crm/services/impl/cost/CostCategoryServiceImpl.java
@@ -1,22 +1,27 @@
 package com.cafe.crm.services.impl.cost;
 
 import com.cafe.crm.exceptions.cost.category.CostCategoryDataException;
+import com.cafe.crm.models.cost.Cost;
 import com.cafe.crm.models.cost.CostCategory;
 import com.cafe.crm.repositories.cost.CostCategoryRepository;
+import com.cafe.crm.repositories.cost.CostRepository;
 import com.cafe.crm.services.interfaces.cost.CostCategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class CostCategoryServiceImpl implements CostCategoryService {
 
 	private final CostCategoryRepository costCategoryRepository;
+	private final CostRepository costService;
 
 	@Autowired
-	public CostCategoryServiceImpl(CostCategoryRepository goodsCategoryRepository) {
+	public CostCategoryServiceImpl(CostCategoryRepository goodsCategoryRepository, CostRepository costService) {
 		this.costCategoryRepository = goodsCategoryRepository;
+		this.costService = costService;
 	}
 
 	@Override
@@ -60,7 +65,21 @@ public class CostCategoryServiceImpl implements CostCategoryService {
 
 	@Override
 	public void delete(Long id) {
-		costCategoryRepository.delete(id);
+		List<Cost> allCostsOnCategory = costService.findByCategoryNameAndVisibleIsTrue(costCategoryRepository.findOne(id).getName());
+		if (allCostsOnCategory == null || allCostsOnCategory.size() == 0) {
+			costCategoryRepository.delete(id);
+		} else {
+			throw new CostCategoryDataException("Остались не удаленные расходы на категории " + printAllCostOnCategory(allCostsOnCategory));
+		}
+	}
+
+	private String printAllCostOnCategory(List<Cost> costs) {
+		if (costs.size() < 3) {
+			return costs.stream().map(Cost::getName).collect(Collectors.joining("\n"));
+		} else {
+			String limitList = costs.stream().limit(2).map(Cost::getName).collect(Collectors.joining("\n"));
+			return limitList + " и др.";
+		}
 	}
 
 	private void checkForUniqueName(CostCategory costCategory) {
